+++ packages/core/src/shared/renderers/css/color.ts
@@ -1,9 +1,9 @@
-import type { ConversionOptions, SimpleCSSProperties } from "../../types/convert.js";
-import type { DTCGColorValue } from "../../types/dtcg-color.js";
-import { convertColorToString } from "../color/color-conversion.js";
-import { isDTCGColorValue } from "../color/color-validation.js";
-import { ErrorMessages } from "../constants/error-messages.js";
-import { isReference } from "../guards.js";
+import type { ConversionOptions, SimpleCSSProperties } from "../../../types/convert.js";
+import type { DTCGColorValue } from "../../../types/dtcg-color.js";
+import { convertColorToString } from "../../color/color-conversion.js";
+import { isDTCGColorValue } from "../../color/color-validation.js";
+import { ErrorMessages } from "../../constants/error-messages.js";
+import { isReference } from "../../guards.js";
 
 export function convertColorToken(
     value: string | DTCGColorValue,
