+++ packages/core/src/shared/renderers/css/cubic-bezier.ts
@@ -1,6 +1,6 @@
-import type { SimpleCSSProperties } from "../../types/convert.js";
-import type { TokenValue } from "../../types/tokens.js";
-import { isReference } from "../guards.js";
+import type { SimpleCSSProperties } from "../../../types/convert.js";
+import type { TokenValue } from "../../../types/tokens.js";
+import { isReference } from "../../guards.js";
 
 export function convertCubicBezierToken(value: TokenValue<"cubicBezier">): SimpleCSSProperties {
     if (isReference(value)) {
