+++ packages/core/src/shared/renderers/css/gradient.ts
@@ -1,9 +1,9 @@
-import type { CSSGradientProperties, ConversionOptions } from "../../types/convert.js";
-import type { GradientStop, TokenValue } from "../../types/tokens.js";
-import { convertColorToString } from "../color/color-conversion.js";
-import { isDTCGColorValue } from "../color/color-validation.js";
-import { ErrorMessages } from "../constants/error-messages.js";
-import { isReference } from "../guards.js";
+import type { CSSGradientProperties, ConversionOptions } from "../../../types/convert.js";
+import type { GradientStop, TokenValue } from "../../../types/tokens.js";
+import { convertColorToString } from "../../color/color-conversion.js";
+import { isDTCGColorValue } from "../../color/color-validation.js";
+import { ErrorMessages } from "../../constants/error-messages.js";
+import { isReference } from "../../guards.js";
 
 function convertGradientColor(color: GradientStop["color"], options: ConversionOptions): string {
     if (isReference(color)) {
