+++ packages/core/tests/converters/color.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { convertColorToken } from "../../src/shared/converters/color.js";
+import { convertColorToken } from "../../src/shared/renderers/css/color.js";
 import type { ConversionOptions } from "../../src/types/convert.js";
 
 const createOptions = (overrides: Partial<ConversionOptions> = {}): ConversionOptions => ({
