+++ packages/core/tests/converters/cubic-bezier.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { convertCubicBezierToken } from "../../src/shared/converters/cubic-bezier.js";
+import { convertCubicBezierToken } from "../../src/shared/renderers/css/cubic-bezier.js";
 
 describe("convertCubicBezier", () => {
     it("should handle reference values", () => {
